requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
requests
jinja2==3.0.3
httpx
bs4
python-dotenv
openai
pandas
openpyxl
werkzeug==2.0.1
aiofiles==0.7.0
sqlmodel==0.0.8
aiosqlite==0.19.0